Move OutOfRegistersException to c.o.g.lir.alloc.
